+++ tests/mock/offers.mock.ts
@@ -1,71 +1,22 @@
-export const offersMock = [
+import type { Offer } from '../../src/gather'
+
+export const offersMock: Offer[] = [
   {
-    affiliateUrl: 'https://www.testurl.com',
-    edition: '1',
-    id: 134977773,
-    isActive: true,
-    merchant: '557',
-    platform: 'PS5',
-    price: {
-      eur: {
-        bestCoupon: {
-          code: 'AKS10',
-          discountStrategy: '%',
-          discountValue: '10.00',
-          isCashback: false,
-        },
-        currency: 'eur',
-        price: 24.74,
-        priceWithoutCoupon: 27.49,
-      },
-    },
-    region: '9',
-    stock: 'InStock',
+    merchant: 'Kinguin',
+    edition: 'Standard Edition',
+    region: 'Steam',
+    currentPrice: 38.66,
+    minDiscountPrice: 37.37,
+    couponCode: 'AKSGAME',
+    lastUpdate: '2026-06-19 18:28:55',
   },
   {
-    affiliateUrl: 'https://www.testurl.com',
-    edition: '1',
-    id: 135023156,
-    isActive: true,
-    merchant: '408',
-    platform: 'PS5',
-    price: {
-      eur: {
-        bestCoupon: {
-          code: '14%: AKS14',
-          discountStrategy: '%',
-          discountValue: '14.00',
-          isCashback: false,
-        },
-        currency: 'eur',
-        price: 25.09,
-        priceWithoutCoupon: 29.17,
-      },
-    },
-    region: '9',
-    stock: 'InStock',
-  },
-  {
-    affiliateUrl: 'https://www.testurl.com',
-    edition: '1',
-    id: 134470740,
-    isActive: true,
-    merchant: '47',
-    platform: 'steam',
-    price: {
-      eur: {
-        bestCoupon: {
-          code: '12%: AKS12',
-          discountStrategy: '%',
-          discountValue: '12.00',
-          isCashback: false,
-        },
-        currency: 'eur',
-        price: 25.71,
-        priceWithoutCoupon: 29.22,
-      },
-    },
-    region: '9',
-    stock: 'InStock',
+    merchant: 'G2A',
+    edition: 'Standard Edition',
+    region: 'Steam',
+    currentPrice: 41.19,
+    minDiscountPrice: 38.52,
+    couponCode: 'AKSHERO',
+    lastUpdate: '2026-06-19 03:02:53',
   },
 ]